requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
requests==2.31.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
